+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Persist `Environment` in metadata ‒ [#1741](https://github.com/paritytech/ink/pull/1741)
 - Added possibility for `runtime_call` in E2E tests ‒ [#1736](https://github.com/paritytech/ink/pull/1736)
-- Added `default` attribute to constructors and messages ‒ [#1703](https://github.com/paritytech/ink/pull/1724)
+- Added `default` attribute to constructors and messages ‒ [#1724](https://github.com/paritytech/ink/pull/1724)
 
 ### Changed
 - Upgraded `syn` to version `2` ‒ [#1731](https://github.com/paritytech/ink/pull/1731)