+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Persist `Environment` in metadata ‒ [#1741](https://github.com/paritytech/ink/pull/1741)
 - Added possibility for `runtime_call` in E2E tests ‒ [#1736](https://github.com/paritytech/ink/pull/1736)
 - Added `default` attribute to constructors and messages ‒ [#1724](https://github.com/paritytech/ink/pull/1724)
+- Added clarification about `Mapping::size` unit ‒ [#1735](https://github.com/paritytech/ink/pull/1735)
 
 ### Changed
 - Upgraded `syn` to version `2` ‒ [#1731](https://github.com/paritytech/ink/pull/1731)